requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pyodbc==5.1.0
reportlab==4.2.0
